pyproject.toml
[tool.poetry]
name = "advanced-techniques"
version = "0.1.0"
description = "Advanced Transformer Techniques - Working Examples"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
transformers = "^4.36.0"
torch = "^2.0.0"
pyarrow = ">=11.0.0,<18.0.0"
datasets = "^2.14.0"
tokenizers = "^0.15.0"
accelerate = "^0.26.0"
evaluate = "^0.4.0"
langdetect = "^1.0.9"
peft = "^0.7.0"
scipy = "^1.11.0"
langchain = "^0.2.0"
openai = "^1.0.0"
anthropic = "^0.20.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.0"
numpy = "^1.26.0"
matplotlib = "^3.10.3"
seaborn = "^0.13.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
ruff = "^0.6.0"
jupyter = "^1.0.0"
ipykernel = "^6.29.0"
matplotlib = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
